Change error handling (#2)
The functions which are called to deal with any errors that crop up were
modified to also get passed the http request. Additionally the
ErrHandler Presenter was modified so if a non-zero Response is returned
from an ErrPresenter that Response gets returned regardless if an error
was also returned (as opposed to before where if an error was returned
then the default response took effect). I thought this made the unit
more flexible.